% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_recipe.R
\name{compute_project_indicators}
\alias{compute_project_indicators}
\title{Calculer tous les indicateurs d'un projet a partir de recettes}
\usage{
compute_project_indicators(project, recipes)
}
\arguments{
\item{project}{Un \code{mecore::me_project}.}

\item{recipes}{Liste de \code{me_indicator_recipe}.}
}
\value{
Le \code{mecore::me_project} mis a jour (nouveaux indicateurs ajoutes).
}
\description{
Pour chaque recette, retrouve le \code{me_dataset} correspondant dans
\code{project@datasets} (par nom), calcule l'indicateur (ou les indicateurs,
si \code{group_by} est renseigne), et l'ajoute a \code{project@indicators}.
}
